+++ libs/base_core/src/code/error/errors.rs
@@ -11,30 +11,28 @@
 // - composite error types:
 //   - NotAvailable: NotImplemented + NotSupported
 
-use crate::{_TAG_NO, define_error};
+use crate::{_TAG_NO, _TAG_VALUE, define_error};
 
 /* individual errors */
 
 define_error![individual: pub struct FailedErrorConversion;
-    +location: "code/error",
+    #[derive(Default)], +location: "code/error", +tag: _TAG_VALUE!(),
     DOC_FAILED_CONVERSION = "A failed conversion between two error types.",
     self+f => write!(f, "Failed to convert between error types."),
 ];
 define_error![individual: pub struct NotImplemented;
-    +location: "code/error",
-    +tag: _TAG_NO!(),
+    #[derive(Default)], +location: "code/error", +tag: _TAG_NO!(),
     DOC_NOT_IMPLEMENTED = "The requested functionality is not implemented.",
     self+f => write!(f, "The requested functionality is not implemented."),
 ];
 define_error![individual: pub struct NotSupported;
-    +location: "code/error",
-    +tag: _TAG_NO!(),
+    #[derive(Default)], +location: "code/error", +tag: _TAG_NO!(),
     DOC_NOT_SUPPORTED = "The requested functionality is not supported by this type.",
     self+f => write!(f, "The requested functionality is not supported by this type."),
 ];
 
 define_error![individual: pub struct InvalidValue;
-    +location: "code/error",
+    #[derive(Default)], +location: "code/error", +tag: _TAG_VALUE!(),
     DOC_INVALID_VALUE = "An invalid value was received for the given type or operation.",
     self+f => write!(f, "An invalid value was received for the given type or operation."),
 ];