+++ libs/base_core/src/num/error.rs
@@ -22,47 +22,40 @@ use crate::{
 /* individual errors */
 
 define_error![individual: pub struct NoInverse;
-    +location: "num/error",
-    +tag: _tags!(num),
+    #[derive(Default)], +location: "num/error", +tag: _tags!(num),
     DOC_NO_INVERSE = "An inverse doesn't exist.",
     self+f => f.write_str(DOC_NO_INVERSE!()),
 ];
 define_error![individual: pub struct MismatchedSizes;
-    +location: "num/error",
-    +tag: _tags!(num),
+    #[derive(Default)], +location: "num/error", +tag: _tags!(num),
     DOC_MISMATCHED_SIZES = "The provided values are not compatible in size.",
     self+f => f.write_str(DOC_MISMATCHED_SIZES!()),
 ];
 define_error![individual: pub struct IncompatibleBounds;
-    +location: "num/error",
-    +tag: _tags!(num),
+    #[derive(Default)], +location: "num/error", +tag: _tags!(num),
     DOC_INCOMPATIBLE_BOUNDS = "The given bounds are incompatible.
 E.g. lower bound exceeds upper bound.", // IMPROVE +docs
     self+f => f.write_str("The given bounds are incompatible."),
 ];
 
 define_error![individual: pub struct NonNegativeRequired;
-    +location: "num/error",
-    +tag: _tags!(num),
+    #[derive(Default)], +location: "num/error", +tag: _tags!(num),
     DOC_NON_NEGATIVE_REQUIRED = "A non-negative value is required.",
     self+f => f.write_str(DOC_NON_NEGATIVE_REQUIRED!()),
 ];
 define_error![individual: pub struct PositiveRequired;
-    +location: "num/error",
-    +tag: _tags!(num),
+    #[derive(Default)], +location: "num/error", +tag: _tags!(num),
     DOC_POSITIVE_REQUIRED = "A positive value is required.",
     self+f => f.write_str(DOC_POSITIVE_REQUIRED!()),
 ];
 define_error![individual: pub struct NonZeroRequired;
-    +location: "num/error",
-    +tag: _tags!(num),
+    #[derive(Default)], +location: "num/error", +tag: _tags!(num),
     DOC_NON_ZERO_REQUIRED = "A non-zero value is required.",
     self+f => f.write_str(DOC_NON_ZERO_REQUIRED!()),
 ];
 
 define_error![individual: pub struct Overflow(pub Option<Sign>);
-    +location: "num/error",
-    +tag: _tags!(num),
+    #[derive(Default)], +location: "num/error", +tag: _tags!(num),
     DOC_OVERFLOW = "An arithmetic overflow error, with an optional associated sign.",
     self+f => if let Some(sign) = self.0 {
         match sign {
@@ -114,7 +107,7 @@ impl IntError {
     // }
 }
 #[doc = crate::_tags!(num result)]
-/// A result of a fallibe integer operation.
+/// A result of a fallible integer operation.
 #[doc = crate::_doc_location!("num")]
 pub type IntResult<T> = crate::Result<T, IntError>;