+++ libs/base_core/src/phys/time/mod.rs
@@ -6,9 +6,10 @@
 mod _reexport; // SYMLINK from /src/phys/time/_reexport_core.rs
 
 mod errors {
-    use crate::define_error;
+    use crate::{_TAG_TIME, define_error};
 
     define_error! { individual: pub struct Timeout;
+        #[derive(Default)], +location: "phys/time", +tag: _TAG_TIME!(),
         DOC_KEY_ALREADY_EXISTS = "The operation has exceeded the allowed execution time.",
         self+f => write!(f, "The operation has exceeded the allowed execution time.")
     }